requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-docx>=0.8.11
streamlit>=1.25.0
gradio>=3.40.0
openai>=1.0.0
langchain>=0.0.300
langchain-community>=0.0.20
chromadb>=0.4.0
tiktoken>=0.5.0
